requirements.txt
django==2.0.4
djangorestframework
markdown
django-filter